refactor: extract media understanding common package
