+++ packages/media-understanding-common/src/active-model.ts
@@ -1,0 +1,4 @@
+export type ActiveMediaModel = {
+  provider: string;
+  model?: string;
+};
