+++ packages/media-understanding-common/src/errors.ts
@@ -1,0 +1,21 @@
+type MediaUnderstandingSkipReason =
+  | "maxBytes"
+  | "timeout"
+  | "unsupported"
+  | "empty"
+  | "blocked"
+  | "tooSmall";
+
+export class MediaUnderstandingSkipError extends Error {
+  readonly reason: MediaUnderstandingSkipReason;
+
+  constructor(reason: MediaUnderstandingSkipReason, message: string) {
+    super(message);
+    this.reason = reason;
+    this.name = "MediaUnderstandingSkipError";
+  }
+}
+
+export function isMediaUnderstandingSkipError(err: unknown): err is MediaUnderstandingSkipError {
+  return err instanceof MediaUnderstandingSkipError;
+}
