+++ scripts/build-all.mjs
@@ -48,6 +48,7 @@ export const BUILD_ALL_STEPS = [
         "packages/plugin-sdk/package.json",
         "packages/llm-core/package.json",
         "packages/markdown-core/package.json",
+        "packages/media-understanding-common/package.json",
         "packages/terminal-core/package.json",
         "packages/memory-host-sdk/package.json",
         "tsconfig.json",
@@ -57,6 +58,7 @@ export const BUILD_ALL_STEPS = [
         "packages/markdown-core/src",
         "packages/memory-host-sdk/src",
         "packages/media-generation-core/src",
+        "packages/media-understanding-common/src",
         "packages/terminal-core/src",
         "src/types",
         "src/video-generation/dashscope-compatible.ts",
