+++ src/media-understanding/active-model.types.ts
@@ -1,4 +1,1 @@
-export type ActiveMediaModel = {
-  provider: string;
-  model?: string;
-};
+export * from "../../packages/media-understanding-common/src/active-model.js";
