+++ src/media-understanding/errors.ts
@@ -1,21 +1,1 @@
-type MediaUnderstandingSkipReason =
-  | "maxBytes"
-  | "timeout"
-  | "unsupported"
-  | "empty"
-  | "blocked"
-  | "tooSmall";
-
-export class MediaUnderstandingSkipError extends Error {
-  readonly reason: MediaUnderstandingSkipReason;
-
-  constructor(reason: MediaUnderstandingSkipReason, message: string) {
-    super(message);
-    this.reason = reason;
-    this.name = "MediaUnderstandingSkipError";
-  }
-}
-
-export function isMediaUnderstandingSkipError(err: unknown): err is MediaUnderstandingSkipError {
-  return err instanceof MediaUnderstandingSkipError;
-}
+export * from "../../packages/media-understanding-common/src/errors.js";
