+++ src/media-understanding/openai-compatible-video.ts
@@ -1,66 +1,1 @@
-import { normalizeOptionalString } from "../shared/string-coerce.js";
-import { normalizeTrimmedStringList } from "../shared/string-normalization.js";
-
-export type OpenAiCompatibleVideoPayload = {
-  choices?: Array<{
-    message?: {
-      content?: string | Array<{ text?: string }>;
-      reasoning_content?: string;
-    };
-  }>;
-};
-
-export function resolveMediaUnderstandingString(
-  value: string | undefined,
-  fallback: string,
-): string {
-  const trimmed = normalizeOptionalString(value);
-  return trimmed || fallback;
-}
-
-export function coerceOpenAiCompatibleVideoText(
-  payload: OpenAiCompatibleVideoPayload,
-): string | null {
-  const message = payload.choices?.[0]?.message;
-  if (!message) {
-    return null;
-  }
-  if (typeof message.content === "string" && message.content.trim()) {
-    return message.content.trim();
-  }
-  if (Array.isArray(message.content)) {
-    const text = normalizeTrimmedStringList(message.content.map((part) => part.text)).join("\n");
-    if (text) {
-      return text;
-    }
-  }
-  if (typeof message.reasoning_content === "string" && message.reasoning_content.trim()) {
-    return message.reasoning_content.trim();
-  }
-  return null;
-}
-
-export function buildOpenAiCompatibleVideoRequestBody(params: {
-  model: string;
-  prompt: string;
-  mime: string;
-  buffer: Buffer;
-}) {
-  return {
-    model: params.model,
-    messages: [
-      {
-        role: "user",
-        content: [
-          { type: "text", text: params.prompt },
-          {
-            type: "video_url",
-            video_url: {
-              url: `data:${params.mime};base64,${params.buffer.toString("base64")}`,
-            },
-          },
-        ],
-      },
-    ],
-  };
-}
+export * from "../../packages/media-understanding-common/src/openai-compatible-video.js";
