+++ src/media-understanding/provider-id.ts
@@ -1,23 +1,1 @@
-import { normalizeProviderId } from "../agents/provider-id.js";
-
-export function normalizeMediaProviderId(id: string): string {
-  const normalized = normalizeProviderId(id);
-  if (normalized === "gemini") {
-    return "google";
-  }
-  if (normalized === "minimax-cn") {
-    return "minimax";
-  }
-  if (normalized === "minimax-portal-cn") {
-    return "minimax-portal";
-  }
-  return normalized;
-}
-
-export function normalizeMediaExecutionProviderId(id: string): string {
-  const normalized = normalizeProviderId(id);
-  if (normalized === "minimax-cn" || normalized === "minimax-portal-cn") {
-    return normalized;
-  }
-  return normalizeMediaProviderId(normalized);
-}
+export * from "../../packages/media-understanding-common/src/provider-id.js";
