+++ src/media-understanding/provider-supports.ts
@@ -1,17 +1,1 @@
-import type { MediaUnderstandingCapability, MediaUnderstandingProvider } from "./types.js";
-
-export function providerSupportsCapability(
-  provider: MediaUnderstandingProvider | undefined,
-  capability: MediaUnderstandingCapability,
-): boolean {
-  if (!provider) {
-    return false;
-  }
-  if (capability === "audio") {
-    return Boolean(provider.transcribeAudio);
-  }
-  if (capability === "image") {
-    return Boolean(provider.describeImage);
-  }
-  return Boolean(provider.describeVideo);
-}
+export * from "../../packages/media-understanding-common/src/provider-supports.js";
