+++ src/media-understanding/video.ts
@@ -1,10 +1,1 @@
-import { DEFAULT_VIDEO_MAX_BASE64_BYTES } from "./defaults.constants.js";
-
-export function estimateBase64Size(bytes: number): number {
-  return Math.ceil(bytes / 3) * 4;
-}
-
-export function resolveVideoMaxBase64Bytes(maxBytes: number): number {
-  const expanded = Math.floor(maxBytes * (4 / 3));
-  return Math.min(expanded, DEFAULT_VIDEO_MAX_BASE64_BYTES);
-}
+export * from "../../packages/media-understanding-common/src/video.js";
